pyproject.toml
[project]
name = "taskflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt>=5.0.0",
    "celery>=5.5.3",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "mysqlclient>=2.2.7",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pydantic[dotenv,email]>=2.12.4",
    "pytest>=8.4.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
